requirements.txt
python-telegram-bot==21.3
ytmusicapi==1.7.0
